requirements.txt
fastapi
uvicorn
requests
python-dotenv
gunicorn
supervisor